feat(html5bp): Docs, assets (js, css), and HTML from HTML5 Boilerplate
This effectivelymerges the entirety of HTML5 Boilerplate in with
this skeleton.

This adds all of HTML5 boilerplate's documentation, the static
resources (browserconfig.xml, site.webmanifest, robots.txt, humans.txt),
as well as all of the icons included (icon.png, favicon.ico, etc) to the
skeleton.

At this point, it is still without any PHP or server-side JS.